+++ client-side/src/cpp/BrowserJS.cpp
@@ -2,7 +2,10 @@
 
 #ifdef Q_OS_WASM
     #include <emscripten.h>
+    #include <emscripten/val.h>
 #endif
+
+BrowserJS* BrowserJS::s_instance = nullptr;
 
 BrowserJS::BrowserJS(QObject *parent)
     :   QObject{parent},
@@ -11,16 +14,39 @@ BrowserJS::BrowserJS(QObject *parent)
 #else
         b_browserEnvironment{false}
 #endif
-{}
+{
+    s_instance = this;
+}
 
-bool BrowserJS::isBrowserEnvionment() const {
+bool BrowserJS::isBrowserEnvironment() const {
     return b_browserEnvironment;
 }
 
-int BrowserJS::runJS(const QString & code) {
+BrowserJS* BrowserJS::instance() {
+    return s_instance;
+}
+
+int BrowserJS::runIntJS(const QString & code) {
 #ifdef Q_OS_WASM
     return emscripten_run_script_int(code.toLatin1().data());
 #else
     return 0;
+#endif
+}
+
+QString BrowserJS::runStringJS(const QString & code) {
+#ifdef Q_OS_WASM
+    char* result = emscripten_run_script_string(code.toLatin1().data());
+    QString qstr = QString::fromUtf8(result);
+    free(result);
+    return qstr;
+#else
+    return QString();
+#endif
+}
+
+void BrowserJS::runVoidJS(const QString & code) {
+#ifdef Q_OS_WASM
+    emscripten_run_script(code.toLatin1().data());
 #endif
 }